Undo 728ca36e in part of indices type for Mesh objects - better keep them as GLuints

+++ src/model/Mesh.h
@@ -34,7 +34,7 @@ struct Texture
 class Mesh
 {
 public:
-  Mesh(std::vector<Vertex> vertices, std::vector<Texture> textures, std::vector<GLushort> indices);
+  Mesh(std::vector<Vertex> vertices, std::vector<Texture> textures, std::vector<GLuint> indices);
   void setupMesh();
   void setupInstances(glm::mat4* models, unsigned int numModels);
   void draw(Shader& shader, const glm::vec2 &cameraPositionXZ, std::vector<ModelChunk>& chunks, unsigned int index,
@@ -43,7 +43,7 @@ public:
 private:
   std::vector<Vertex> vertices;
   std::vector<Texture> textures;
-  std::vector<GLushort> indices;
+  std::vector<GLuint> indices;
   unsigned int VBO, EBO, instanceVBO, multiDE_I_DIBO;
   unsigned int numInstances;
 };
